refactor(OutputTest): delete writeOut_TableDefinitionEvent
